requirements.txt
Django==2.2.6
gunicorn==19.9.0
psycopg2-binary==2.8.3
django-jet
shodan
django_compressor
django-material-widgets
djangorestframework
django-allauth